requirements.txt
django
django_hide
Pillow
whitenoise
gunicorn
psycopg2-binary
django-quill-editor
django-recaptcha
validate-email-address
py3dns
unidecode